# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: SumUp-inspired payments platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/transactions:
    get:
      operationId: listTransactions
      summary: List transactions
      tags: [payments]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: start_date
          schema:
            type: string
            format: date
        - in: query
          name: end_date
          schema:
            type: string
            format: date
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, successful, failed, refunded]
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"

    post:
      operationId: createTransaction
      summary: Create a simulated transaction
      tags: [payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /payments/transactions/{transactionId}:
    get:
      operationId: getTransaction
      summary: Get a transaction
      tags: [payments]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: transactionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/transactions/{transactionId}/refund:
    post:
      operationId: refundTransaction
      summary: Refund a transaction
      tags: [payments]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: transactionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefundRequest"
      responses:
        "200":
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /payments/checkouts:
    post:
      operationId: createCheckout
      summary: Create a payment link
      tags: [payments]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutRequest"
      responses:
        "201":
          description: Checkout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkout"

    get:
      operationId: listCheckouts
      summary: List payment links
      tags: [payments]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Checkout list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkout"
  /payments/summary:
    get:
      operationId: getPaymentSummary
      summary: Get aggregated payment stats
      tags: [payments]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [today, week, month, year]
            default: month
      responses:
        "200":
          description: Payment summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSummary"
  /payments/payouts:
    get:
      operationId: listPayouts
      summary: List payouts
      tags: [payments]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payout list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payout"

  # ─── PRODUCTS ─────────────────────────────────────────────────────────────
  /products:
    get:
      operationId: listProducts
      summary: List products
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: active
          schema:
            type: boolean
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      summary: Create a product
      tags: [products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      summary: Get a product
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProduct
      summary: Update a product
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      summary: Delete a product
      tags: [products]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /products/categories:
    get:
      operationId: listProductCategories
      summary: List product categories
      tags: [products]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductCategory"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    Transaction:
      type: object
      required: [id, merchant_id, amount, currency, status, payment_method, created_at]
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, successful, failed, refunded, partially_refunded]
        payment_method:
          type: string
          enum: [card, tap_to_pay, payment_link, cash]
        card_last_four:
          type: string
        card_scheme:
          type: string
          enum: [visa, mastercard, amex, maestro]
        product_id:
          type: string
        product_name:
          type: string
        description:
          type: string
        receipt_sent:
          type: boolean
        customer_email:
          type: string
        created_at:
          type: string
          format: date-time
        refunded_at:
          type: string
          format: date-time
    TransactionListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    CreateTransactionRequest:
      type: object
      required: [amount, currency, payment_method]
      properties:
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
        payment_method:
          type: string
          enum: [card, tap_to_pay, payment_link, cash]
        description:
          type: string
        product_id:
          type: string
        customer_email:
          type: string
    RefundRequest:
      type: object
      properties:
        amount:
          type: number
    CreateCheckoutRequest:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
        currency:
          type: string
        title:
          type: string
        description:
          type: string
        redirect_url:
          type: string
    Checkout:
      type: object
      required: [id, merchant_id, amount, currency, status, checkout_url, created_at]
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, paid, expired, cancelled]
        checkout_url:
          type: string
        redirect_url:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    PaymentSummary:
      type: object
      required: [period, total_volume, total_transactions, currency, daily_breakdown]
      properties:
        period:
          type: string
        total_volume:
          type: number
        total_transactions:
          type: integer
        average_transaction:
          type: number
        currency:
          type: string
        daily_breakdown:
          type: array
          items:
            $ref: "#/components/schemas/DailyVolume"
        by_payment_method:
          type: array
          items:
            $ref: "#/components/schemas/MethodVolume"
    DailyVolume:
      type: object
      required: [date, volume, count]
      properties:
        date:
          type: string
          format: date
        volume:
          type: number
        count:
          type: integer
    MethodVolume:
      type: object
      required: [method, volume, count, percentage]
      properties:
        method:
          type: string
        volume:
          type: number
        count:
          type: integer
        percentage:
          type: number
    Payout:
      type: object
      required: [id, merchant_id, amount, currency, status, payout_date, created_at]
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, processing, paid, failed]
        payout_date:
          type: string
          format: date
        reference:
          type: string
        created_at:
          type: string
          format: date-time

    # ─── PRODUCTS ────────────────────────────────────────────────────────────
    Product:
      type: object
      required: [id, merchant_id, name, price, currency, active, created_at]
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        category:
          type: string
        sku:
          type: string
        active:
          type: boolean
        image_url:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProductRequest:
      type: object
      required: [name, price, currency]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          type: string
        category:
          type: string
        sku:
          type: string
        active:
          type: boolean
          default: true
        image_url:
          type: string
    ProductCategory:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        merchant_id:
          type: string