# lib/api-zod/src/generated/types/paymentSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SumUp-inspired payments platform API
 * OpenAPI spec version: 1.0.0
 */
import type { DailyVolume } from "./dailyVolume";
import type { MethodVolume } from "./methodVolume";

export interface PaymentSummary {
  period: string;
  total_volume: number;
  total_transactions: number;
  average_transaction?: number;
  currency: string;
  daily_breakdown: DailyVolume[];
  by_payment_method?: MethodVolume[];
}