# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SumUp-inspired payments platform API
 * OpenAPI spec version: 1.0.0
 */

export interface Product {
  id: string;
  merchant_id: string;
  name: string;
  description?: string;
  price: number;
  currency: string;
  category?: string;
  sku?: string;
  active: boolean;
  image_url?: string;
  created_at: Date;
  updated_at?: Date;
}